update after review
- integrate suggested changes
